run swaylock as a systemd service
this ensures that even if swaylock crashes, it'll be restarted. inspired
after a real-world crash of swaylock which left the screen unlocked.

`Type=forking` and `swaylock -f` is used so that the locked state is
propagated as a ready service, thus blocking `systemctl start` until the
screen is really locked. In combination with `swayidle -w` that ensures
that swayidle releases suspend inhibitors only after the screen is
really locked.